Refactored: Splitted Bus.java into Menu.java so that the contribution is easier
